Refactor youtubeApiExample.py to youtube_api.py for consistency
